requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiosqlite==0.20.0
    # via tortoise-orm
alembic==1.14.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
asyncpg==0.30.0
    # via -r requirements.in
django==5.1.3
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
iso8601==2.1.0
    # via tortoise-orm
mako==1.3.6
    # via alembic
markupsafe==3.0.2
    # via mako
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.9.2
    # via tortoise-orm
pydantic-core==2.23.4
    # via pydantic
pypika-tortoise==0.2.1
    # via tortoise-orm
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via tortoise-orm
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   alembic
sqlparse==0.5.1
    # via django
tortoise-orm==0.21.7
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   aiosqlite
    #   alembic
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.2
    # via django
